% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_comhis.R
\name{theme_comhis}
\alias{theme_comhis}
\title{Theme (ggplot) for Helsinki Computational History group}
\usage{
theme_comhis(type = "invalid", base_size = 20, family = "Helvetica")
}
\arguments{
\item{type}{Plot type ("continuous"; "discrete").}

\item{base_size}{Base size for fonts.}

\item{family}{Font family.}
}
\description{
Default ggplot theme of Helsinki Computational History group
}
\examples{
  
  #library(ggplot2)
  #data(iris)
  #p <- ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length)) +
  #  geom_point()
  #p + theme_comhis("continuous")

}
\author{
Abhishekh Gupta, Ville Vaara, Leo Lahti.
}
